pyproject.toml
[project]
name = "smartseason"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework-simplejwt>=5.5.1",
]
